+++ src/models/components/atoms/Avatar.ts
