+++ src/models/components/atoms/BottomSheet.ts
@@ -1,0 +1,8 @@
+import { ReactNode, RefObject } from 'react'
+
+export type BottomSheetPropsType = {
+  children?: ReactNode
+  isexpanded: boolean
+  bottomSheetRef: RefObject<HTMLDivElement>
+  filterType?: string
+}
